+++ NEWS.md
@@ -1,4 +1,10 @@
 # QCkit v1.0.1 (development version)
+2025-03-11
+* Updated function `get_park_names()` to include user-supplied missing values, identify unit codes that map to no park names and multiple park names
+
+2025-03-10
+* Added function `get_park_names()`, which adds a park name column to a data frame that has a unit code column, using a public IRMA API
+
 2025-02-25
 * Updated `CONTRIBUTING.md`.